requirements.txt
python-telegram-bot==20.8
requests==2.31.0
pytz==2024.1
APScheduler==3.10.4 